+++ Apizr/Tests/Apizr.Tests/ApizrTests.cs
@@ -316,6 +316,53 @@ namespace Apizr.Tests
         }
 
         [Fact]
+        public async Task Calling_WithAkavacheCacheHandler_With_Multiple_CacheKey_Should_Cache_Result()
+        {
+            var reqResManager = ApizrBuilder.Current.CreateManagerFor<IReqResUserService>(options =>
+                options.WithLoggerFactory(LoggerFactory.Create(builder =>
+                        builder.AddXUnit(_outputHelper)
+                            .SetMinimumLevel(LogLevel.Trace)))
+                    .WithLogging()
+                    .WithAkavacheCacheHandler()
+                    .WithDelegatingHandler(new TestRequestHandler()));
+
+            // Clearing cache
+            await reqResManager.ClearCacheAsync();
+
+            var testDictionary = new Dictionary<string, object>
+            {
+                { "test1", "test1" },
+                { "test2", 2 }
+            };
+
+            var customTypeParam = new ReadAllUsersParams("test1", 2);
+
+            // Defining a throwing request
+            Func<Task> act1 = () => reqResManager.ExecuteAsync(api => api.GetUserAsync(1, testDictionary, customTypeParam, HttpStatusCode.BadRequest));
+
+            // Calling it at first execution should throw as expected without any cached result
+            var ex = await act1.Should().ThrowAsync<ApizrException<UserDetails>>();
+            ex.And.CachedResult.Should().BeNull();
+
+            // This one should succeed
+            var result = await reqResManager.ExecuteAsync(api => api.GetUserAsync(1, testDictionary, customTypeParam, HttpStatusCode.OK));
+
+            // and cache result in-memory
+            result.Should().NotBeNull();
+
+            // This one should fail but with cached result
+            var ex2 = await act1.Should().ThrowAsync<ApizrException<UserDetails>>();
+            ex2.And.CachedResult.Should().NotBeNull();
+
+            // Defining another throwing request
+            Func<Task> act2 = () => reqResManager.ExecuteAsync(api => api.GetUserAsync(2, testDictionary, customTypeParam, HttpStatusCode.BadRequest));
+
+            // Calling it again with another cache key value should throw as expected but without any cached result
+            var ex3 = await act2.Should().ThrowAsync<ApizrException<UserDetails>>();
+            ex3.And.CachedResult.Should().BeNull();
+        }
+
+        [Fact]
         public async Task RequestTimeout_Should_Be_Handled_By_Polly()
         {
             var maxRetryAttempts = 3;